improve updateCryptoPrice to update price in only one insertion

+++ backend/src/core/cache.go
@@ -8,7 +8,7 @@ import (
 
 type CachePrice struct {
 	KeyName    string
-	Price      int
+	Price      int64
 	LastUpdate time.Time
 }
